crawler/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
global-land-mask==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.10.0
numpy==2.4.4
pillow==12.1.1
playwright==1.58.0
prisma==0.15.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
PyMuPDF==1.27.2.2
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
soupsieve==2.8.3
tomlkit==0.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
